+++ internal/core/config_test.go
@@ -39,6 +39,10 @@ func TestParseDefaultConfig(t *testing.T) {
 				LinkFormat:        "markdown",
 				LinkEncodePath:    true,
 				LinkDropExtension: true,
+				Frontmatter: YamlFrontmatterConfig{
+					CreationDate:     "date",
+					ModificationDate: "modified",
+				},
 			},
 		},
 		Tool: ToolConfig{
@@ -91,6 +95,10 @@ func TestParseComplete(t *testing.T) {
 		link-format = "custom"
 		link-encode-path = true
 		link-drop-extension = false
+
+		[format.markdown.frontmatter]
+		creation-date-key = "created"
+		modification-date-key = "changed"
 
 		[tool]
 		editor = "vim"
@@ -237,6 +245,10 @@ func TestParseComplete(t *testing.T) {
 				LinkFormat:        "custom",
 				LinkEncodePath:    true,
 				LinkDropExtension: false,
+				Frontmatter: YamlFrontmatterConfig{
+					CreationDate:     "created",
+					ModificationDate: "changed",
+				},
 			},
 		},
 		Tool: ToolConfig{
@@ -425,6 +437,10 @@ func TestParseMergesGroupConfig(t *testing.T) {
 				LinkFormat:        "markdown",
 				LinkEncodePath:    true,
 				LinkDropExtension: true,
+				Frontmatter: YamlFrontmatterConfig{
+					CreationDate:     "date",
+					ModificationDate: "modified",
+				},
 			},
 		},
 		LSP: LSPConfig{