+++ internal/core/notebook_store.go
@@ -287,6 +287,11 @@ multiword-tags = true
 multiword-tags = false
 {{/if}}
 
+#[format.markdown.frontmatter]
+
+# Define custom keys and properties of the frontmatter block
+#creation-date-key = "created" # default is "date"
+#modification-date-key = "changed" # default is "modified"
 
 # EXTERNAL TOOLS
 [tool]